requirements.txt
streamlit
openai
python-dotenv
requests
pandas
scikit-learn
pypdf
docxtpl
docx2pdf